modules/packetio: use separate mempools for tx/rx
lwip holds Tx pbuf/mbufs until they are acked which can
consume a lot of buffers when there are multiple open
TCP sessions.

lwip also holds Rx pbuf/mbufs when packets are received out of
sequence, but this is a lot less likely.

When using a single mempool for both tx and rx, if the mempool
is exhuasted it prevents any packet from being received so
the tx packets don't get freed because ack is not received.

Using separate pools for tx/rx prevents unacked tx packets from
blocking rx.

This change also adds configuration options for adding extra
pbuf/mbufs to help with scaling and controlling the size of
the buffers for larger MTUs.