socket: fix performance when no rx interrupt and no sinks
The worker non-interrupt loop was was using a 0 msec idle timeout when
sinks were used and a 10 msec timeout when sinks were not used.

Since there is no interrupt to wakeup the rx thread, the
worker checks the rx queue every 10 msec which is too slow
causing packet loss and TCP performance issues.

This change adds counters for the number of sockets using a port
and modifies the worker to also use a 0 msec timeout when there are
any sockets associated with the port.

The default poll timeout was also reduced from 10 msec to 1 msec
so that the poll timeout change happens faster.