wip: experimenting with aggregating cpu avaiable stats
This change was used to try and figure out how the cpu
available stats is aggregated.

This change tries to aggregate the available time as a
summation instead of a range union.

Looking at places in code which call the sum_stats() function,
it looks like the src/modules/cpu/generator/coordinator.cpp
cooordinator::do_load_update() might need to be updated to
multiple the load by the number of shards/cores.