modules/packetio: NUMA memory missmatch handling
Improve behavior when NUMA memory request does not match
what is available/supported.

The system should be configured so this does not occur,
but that may not always be possible.  Try to prevent errors
if possible and log warnings.

mempool allocation will try to use any available NUMA node if
memory is not available on the requested node.

device numa_node is set to unknown (-1) when NUMA memory is not
allowed on the NUMA node.